+++ apps/dine-pleiepenger/src/i18n/nb.ts
@@ -1,0 +1,144 @@
+export const nb = {
+    'velgSak.dokumentTittel': 'Dine pleiepenger - velg sak',
+    'velgSak.tittel': 'Dine pleiepengesaker',
+    'velgSak.barn.fdato': 'Født: {dato}',
+    'forside.dokumentTittel': 'Dine pleiepenger',
+
+    'snarveier.endring.tittel': 'Meld fra om endring',
+    'snarveier.endring.tekst': 'Ved endringer i arbeidstid og ferie i din pleiepengeperiode.',
+    'snarveier.ettersend.tittel': 'Ettersend dokumenter',
+    'snarveier.ettersend.tekst': 'Ettersendt legeerklæring eller andre dokumenter til din sak.',
+    'snarveier.skrivTilOss.tittel': 'Skriv til oss',
+    'snarveier.skrivTilOss.tekst': 'Har du spørsmål til saken din eller vil melde fra om annet enn jobb og ferie?',
+    'snarveier.nySøknad.tittel': 'Ny pleiepengesøknad',
+    'snarveier.nySøknad.tekst': 'Ved forlengelser, nye arbeidsforhold eller større endringer i perioden.',
+
+    'statusTag.status.underBehandling': 'Til behandling',
+    'statusTag.status.ferdigBehandlet': 'Ferdig behandlet',
+    'statusTag.venteårsak.inntektsmelding': 'Venter på inntektsmelding',
+    'statusTag.venteårsak.legeerklæring': 'Venter på legeerklæring',
+    'statusTag.venteårsak.meldekort': 'Venter på meldekort',
+    'statusTag.venteårsak.søktForTidlig': 'Venter på behandling',
+
+    'sakstype.PP_SYKT_BARN': 'Søknad om pleiepenger sykt barn',
+    'sakstype.PP_ETTERSENDELSE': 'Ettersendelse av dokumenter - Pleiepenger sykt barn',
+    'sakstype.PP_SYKT_BARN_ENDRINGSMELDING': 'Endringsmelding pleiepenger',
+
+    'svarfrist.tittel': 'Saksbehandlingstid',
+    'svarfrist.forventetBehandlingstid':
+        'Forventet behandlingstid er {saksbehandlingstidUker} uker fra vi fikk søknaden din.',
+    'svarfrist.lesMerLenke': 'Se saksbehandlingstider',
+    'svarfrist.dokumenterManglerFrist':
+        'Fordi vi mangler dokumenter kan saksbehandlingstiden bli lenger enn <dato>{frist}</dato>.',
+    'svarfrist.generellFrist': 'Du kan forvente svar innen <dato>{frist}</dato>.',
+
+    'hvaSkjerInfo.tittel': 'Hva skjer etter at du har sendt søknad til oss?',
+    'hvaSkjerInfo.innsent.endringsmelding.tittel': 'Saksgang etter innsendt endringsmelding',
+
+    'hvaSkjerInfo.legeerklæring.tittel': 'Legeerklæring',
+    'hvaSkjerInfo.legeerklæring':
+        'Vi sjekker om vi har fått legeerklæring fra deg. Hvis du ikke sendte den sammen med søknaden din, må du ettersende den til oss så snart du kan.',
+
+    'hvaSkjerInfo.inntektsmelding.tittel': 'Inntektsmelding fra arbeidsgiver',
+    'hvaSkjerInfo.inntektsmelding.avsnitt.1':
+        'Hvis du er arbeidstaker, ser vi om arbeidsgiver har sendt oss inntektsmelding. Den trenger vi hvis du søker for første gang, eller det er mer enn fire uker opphold i perioden med pleiepenger.',
+    'hvaSkjerInfo.inntektsmelding.avsnitt.2.1':
+        'Kontakt arbeidsgiveren din for å sikre at de sender inntektsmeldingen til oss. Du kan henvise arbeidsgiver til ',
+    'hvaSkjerInfo.inntektsmelding.avsnitt.2.2': 'nav.no/arbeidsgiver/pleiepenger-barn',
+    'hvaSkjerInfo.inntektsmelding.avsnitt.2.3': ' hvis de trenger mer informasjon.',
+
+    'hvaSkjerInfo.behandling.tittel': 'Vi behandler søknaden din',
+    'hvaSkjerInfo.behandling':
+        'Når vi har fått legeerklæring og inntektsmelding kan vi behandle søknaden din. Vi kontakter deg hvis vi trenger flere opplysninger fra deg.',
+
+    'hvaSkjerInfo.ferdigBehandlet.tittel': 'Når søknaden er ferdig behandlet',
+    'hvaSkjerInfo.ferdigBehandlet.1': 'Når vi har behandlet søknaden din, får du svar i ',
+    'hvaSkjerInfo.ferdigBehandlet.2': 'dokumentoversikten på Min side',
+    'hvaSkjerInfo.ferdigBehandlet.3': '.',
+
+    'hvaSkjerInfo.utbetaling.tittel': 'Utbetaling',
+    'hvaSkjerInfo.utbetaling.1':
+        'Hvis du skal få pleiepengene direkte fra NAV, utbetaler vi innen den 25. i måneden. Hvis du skal ha etterbetalt pleiepenger for måneder bakover i tid, er de vanligvis på konto innen 5 virkedager fra du fikk svar på søknaden. Du finner ',
+    'hvaSkjerInfo.utbetaling.2': 'oversikt over utbetalingene dine her',
+    'hvaSkjerInfo.utbetaling.3': '.',
+
+    'påbegyntSøknad.info.title': 'Gå til påbegynt søknad',
+    'påbegyntSøknad.info': 'Lagres til og med {datoNårSlettes}',
+    'påbegyntEndring.info.title': 'Gå til påbegynt endringsmelding',
+    'påbegyntEndring.info': 'Lagres til og med {datoNårSlettes}',
+
+    'dineInnsendteSøknader.tittel': 'Dine søknader, endringer og ettersendelser',
+    'innsendtSøknadListe.ingenSøknader': 'Ingen søknader funnet',
+    'innsendtSøknadListe.visFlereInnsendinger': 'Vis flere innsendinger',
+    'innsendtSøknadTitle.mottatt': 'Mottatt: {dato}',
+
+    'dokumenterTittel.PP_SYKT_BARN': 'Søknad og eventuelle vedlegg i søknad',
+    'dokumenterTittel.SØKNAD': 'Søknad og eventuelle vedlegg i søknad',
+    'dokumenterTittel.PP_ETTERSENDELSE': 'Ettersendelse og dokumenter',
+    'dokumenterTittel.ETTERSENDELSE': 'Ettersendelse og dokumenter',
+    'dokumenterTittel.PP_SYKT_BARN_ENDRINGSMELDING': 'Endringsmelding',
+    'dokumenterTittel.ENDRINGSMELDING': 'Endringsmelding',
+    'dokumenter.ingenDokumenter': 'Dokumenter ikke tilgjengelig.',
+    'bekreftelseTilArbeidsgiver.title': 'Bekreftelse til arbeidsgiver',
+    'bekreftelseTilArbeidsgiver.info':
+        'Dette er en bekreftelse som du kan gi til arbeidsgiver. Den bekrefter at vi har fått søknad om pleiepenger fra deg, hvilken periode du har søkt pleiepenger og informasjon om når arbeidsgiver må sende inntektsmelding.',
+    'dokumenterSomKanLastesNed.tittel': 'Dokumenter som kan lastes ned',
+    'dokumenterSomKanLastesNed.bekreftelse': 'Bekreftelse til {organisasjonsnavn}',
+
+    'kontaktOss.tittel': 'Finner du ikke svaret her? Ta kontakt med oss',
+    'kontaktOss.chatLenke': 'Finner du ikke svaret her? Ta kontakt med oss',
+    'kontaktOss.frida':
+        'Du møter først chatbot Frida som har døgnåpent. Mellom klokken 9 og 15 på hverdager kan du be Frida om å få chatte med en veileder.',
+    'kontaktOss.skrivTilOssLenke': 'Skriv til oss',
+    'kontaktOss.skrivTilOssInfo':
+        '<p>Send beskjed eller nye opplysninger i saken din. Du kan også sende spørsmål.</p><p>Svartid er 4 arbeidsdager. Hvis du vil ha svar raskere, kan du bruke chat.</p>',
+    'kontaktOss.ringOssLenke': 'Ring oss på 55 55 33 33',
+    'kontaktOss.ringOssInfo': 'Åpent hverdager kl. 9–15. Vi kan ringe deg tilbake hvis ventetiden er over 5 min.',
+    'kontaktOss.ringOssSeTelefonnummer': 'Se flere telefonnummer og tastevalg',
+
+    'sakPageHeader.saksnr': 'Saksnummer: {saksnr}',
+    'sakPageHeader.pleietrengende': 'Pleietrengende: {navn}',
+    'sakPageHeader.defaultTittel': 'Din pleiepengesak',
+
+    'pdfLenke.lenke': '{tittel} (PDF)',
+
+    'snarveierSak.tittel': 'Snarveier',
+    'snarveierSak.dokumentarkiv': 'Dokumentarkiv',
+    'snarveierSak.utbetalinger': 'Utbetalinger',
+    'snarveierSak.omPleiepenger': 'Om pleiepenger',
+
+    'statusISak.mottattSøknad.tittel': 'Vi har fått søknaden din om pleiepenger',
+    'statusISak.mottattEndringsmelding.tittel': 'Vi har fått endringsmelding fra deg',
+    'statusISak.ferdigBehandlet.tittel': 'Du har fått vedtak om pleiepenger',
+    'statusISak.forventetSvar.søknad.tittel': 'Søknaden er ferdig behandlet',
+    'statusISak.forventetSvar.søknad.info':
+        'Inntektsmelding fra arbeidsgiver og legeerklæring må være sendt inn for at vi kan behandle saken.',
+    'statusISak.forventetSvar.endring.tittel': 'Endringen er ferdig behandlet',
+    'statusISak.forventetSvar.endring.info':
+        'Endringsmeldingen vil oppdatere saken og du vil få ett nytt vedtaksbrev når endringene er ferdig behandlet.',
+    'statusISak.ingenHendelser':
+        '<p>Hei, vi kan dessverre ikke vise informasjon om saken din enda.</p><p>Hvis du har spørsmål, kan du ta kontakt via {lenke}.</p>',
+    'statusISak.ingenHendelser.skrivTilOssLenkeTekst': 'Skriv til oss',
+
+    'venteårsakMelding.legeerklæring.tittel': 'Vi mangler legeerklæring',
+    'venteårsakMelding.legeerklæring.lastOppHer': 'Last opp legeerklæring her',
+    'venteårsakMelding.legeerklæring.info':
+        'Vi har sendt deg brev om at vi trenger flere medisinske opplysninger for å vurdere retten til pleiepenger. Du finner informasjon om hva vi mangler i brevet, <lenke>se brev i dokumentarkivet</lenke>.',
+
+    'venteårsakMelding.inntektsmelding.tittel': 'Vi mangler inntektsmelding fra en eller flere arbeidsgivere',
+    'venteårsakMelding.inntektsmelding.info':
+        'Vi har sendt brev til deg og arbeidsgiveren din, fordi vi mangler inntektsmelding. For å behandle søknaden og beregne hvor mye pleiepenger du kan få, må arbeidsgiver sende inntektsmelding.',
+    'venteårsakMelding.inntektsmelding.dokumentarkivLenke': '<lenke>Se brev i dokumentarkivet</lenke>',
+
+    'venteårsakMelding.søktForTidlig.tittel': 'Vi kan ikke behandle søknaden din enda',
+    'venteårsakMelding.søktForTidlig.info':
+        'For å beregne hvor mye pleiepenger du kan få, trenger vi ferske opplysninger om inntekten din. Arbeidsgiver kan derfor sende inntektsopplysninger tidligst 4 uker før pleiepengeperioden din starter. Saken er satt på vent frem til søknaden kan behandles.',
+
+    'venteårsakMelding.meldekort.tittel': 'Vi venter på at du sender meldekort',
+    'venteårsakMelding.meldekort.info':
+        'Vi kan tidligst behandle søknaden din når du har sendt inn det siste meldekortet, før du starter perioden med pleiepenger.',
+
+    'statusISak.søknadStatusContent.readMoreHeader': 'Se dokumenter og bekreftelse til arbeidsgiver',
+    'statusISak.søknadStatusContent.readMoreHeader.ingenArbeidsgiver': 'Se dokumenter',
+    'statusISak.søknadStatusContent.dokumenterISøknad': 'Søknad og eventuelle vedlegg i søknad',
+};