+++ apps/dine-pleiepenger/src/pages/api/innsynsdata.api.ts
@@ -3,34 +3,46 @@ import { createChildLogger } from '@navikt/next-logger';
 import { HttpStatusCode } from 'axios';
 import { withAuthenticatedApi } from '../../auth/withAuthentication';
 import {
-    fetchSaksbehandlingstid,
     fetchMellomlagringer,
+    fetchSaker,
+    fetchSaksbehandlingstid,
     fetchSøker,
     fetchSøknader,
-    fetchSaker,
 } from '../../server/apiService';
 import { Innsynsdata } from '../../types/InnsynData';
+import { getBrukerprofil } from '../../utils/amplitude/getBrukerprofil';
 import { getXRequestId } from '../../utils/apiUtils';
-import { sortSøknadEtterOpprettetDato } from '../../utils/søknadUtils';
 import { Feature } from '../../utils/features';
-import { getBrukerprofil } from '../../utils/brukerprofilUtils';
+import { sortInnsendtSøknadEtterOpprettetDato } from '../../utils/innsendtSøknadUtils';
+import { fetchAppStatus } from './appStatus.api';
+import dayjs from 'dayjs';
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
     const childLogger = createChildLogger(getXRequestId(req));
-    childLogger.info(`Henter innsynsdata`);
+    childLogger.info(
+        {
+            mellomlagring: Feature.HENT_MELLOMLAGRING,
+            saker: Feature.HENT_SAKER,
+            behandlingstid: Feature.HENT_BEHANDLINGSTID,
+            appstatus: Feature.HENT_APPSTATUS,
+        },
+        `Henter innsynsdata`,
+    );
     try {
         /** Hent søker først for å se om bruker har tilgang */
         const søker = await fetchSøker(req);
 
         /** Bruker har tilgang, hent resten av informasjonen */
-        const [søknaderReq, mellomlagringReq, sakerReq, saksbehandlingstidReq] = await Promise.allSettled([
+        const [søknaderReq, mellomlagringReq, sakerReq, saksbehandlingstidReq, appStatus] = await Promise.allSettled([
             fetchSøknader(req),
-            fetchMellomlagringer(req),
+            Feature.HENT_MELLOMLAGRING ? fetchMellomlagringer(req) : Promise.resolve({}),
             Feature.HENT_SAKER ? fetchSaker(req) : Promise.resolve([]),
             Feature.HENT_BEHANDLINGSTID
                 ? fetchSaksbehandlingstid(req)
                 : Promise.resolve({ saksbehandlingstidUker: undefined }),
+            Feature.HENT_APPSTATUS ? fetchAppStatus() : Promise.resolve(undefined),
         ]);
+        childLogger.info(`Hentet innsynsdata`);
 
         if (søknaderReq.status === 'rejected') {
             childLogger.error(
@@ -38,19 +50,40 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
             );
         }
 
+        childLogger.info(`Parser innsynsdata`);
+
         const saker = sakerReq.status === 'fulfilled' ? sakerReq.value : [];
 
-        const søknader = søknaderReq.status === 'fulfilled' ? søknaderReq.value.sort(sortSøknadEtterOpprettetDato) : [];
+        const innsendteSøknader =
+            søknaderReq.status === 'fulfilled' ? søknaderReq.value.sort(sortInnsendtSøknadEtterOpprettetDato) : [];
+
         const saksbehandlingstidUker =
             saksbehandlingstidReq.status === 'fulfilled'
                 ? saksbehandlingstidReq.value.saksbehandlingstidUker
                 : undefined;
 
-        childLogger.info(getBrukerprofil(søknader, saker, saksbehandlingstidUker), `Hentet innsynsdata`);
+        const brukerprofil = getBrukerprofil(innsendteSøknader, saker, saksbehandlingstidUker);
+        childLogger.info(brukerprofil, `Innsynsdata parset`);
+
+        if (brukerprofil.antallSaker === 0 && brukerprofil.antallSøknader > 0) {
+            const søknadEldreEnnToDager = innsendteSøknader.find((søknad) =>
+                dayjs(søknad.opprettet).isBefore(dayjs().subtract(2, 'days')),
+            );
+            if (søknadEldreEnnToDager) {
+                childLogger.info(
+                    {
+                        ...brukerprofil,
+                        journalpostId: søknadEldreEnnToDager.journalpostId,
+                    },
+                    'Bruker har søknad, men ingen sak',
+                );
+            }
+        }
 
         const innsynsdata: Innsynsdata = {
+            appStatus: appStatus.status === 'fulfilled' ? appStatus.value : undefined,
             søker,
-            søknader,
+            innsendteSøknader,
             mellomlagring: mellomlagringReq.status === 'fulfilled' ? mellomlagringReq.value : {},
             saksbehandlingstidUker,
             saker,