+++ apps/dine-pleiepenger/src/pages/unavailable.page.tsx
@@ -2,8 +2,11 @@ import type { NextPage } from 'next';
 import { BodyShort, GuidePanel, Heading } from '@navikt/ds-react';
 import Head from 'next/head';
 import EmptyPage from '../components/page-layout/empty-page/EmptyPage';
+import { useLogSidevisning } from '@navikt/sif-common-amplitude';
+import { PageKey } from '../types/PageKey';
 
 const UnavailablePage: NextPage = () => {
+    useLogSidevisning(PageKey.utilgjengelig);
     return (
         <EmptyPage>
             <GuidePanel poster={true} className="sm:max-w-lg m-auto">