+++ apps/dine-pleiepenger/src/server/api-models/Behandlingsstatus.ts
@@ -1,0 +1,6 @@
+export enum Behandlingsstatus {
+    OPPRETTET = 'OPPRETTET',
+    UNDER_BEHANDLING = 'UNDER_BEHANDLING',
+    PÅ_VENT = 'PÅ_VENT',
+    AVSLUTTET = 'AVSLUTTET',
+}