+++ apps/dine-pleiepenger/src/server/api-models/Søknadstype.ts
@@ -1,0 +1,6 @@
+export enum Søknadstype {
+    SØKNAD = 'SØKNAD',
+    ENDRINGSMELDING = 'ENDRINGSMELDING',
+    UKJENT = 'UKJENT',
+    // ETTERSENDELSE = 'ETTERSENDELSE',
+}