+++ apps/dine-pleiepenger/src/types/Arbeidsgivere.ts
@@ -1,4 +1,0 @@
-import { z } from 'zod';
-import { ArbeidsgivereSchema } from '../server/api-models/ArbeidsgivereSchema';
-
-export type Arbeidsgivere = z.infer<typeof ArbeidsgivereSchema>;